requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
langchain==0.1.0
langchain-text-splitters==0.0.1
matplotlib==3.8.2
seaborn==0.13.1
pandas==2.1.4
loguru==0.7.2
